Feature(admin-role): created CheckAdmin middleware

+++ routes/api.php
@@ -9,6 +9,7 @@ use App\Http\Controllers\api\ProductController;
 use App\Http\Controllers\roomcontroller;
 use App\Http\Controllers\Api\admincontroller;
 use App\Http\Controllers\imagecontroller;
+use App\Http\Middleware\CheckAdmin;
 
 /*
 |--------------------------------------------------------------------------
@@ -23,17 +24,17 @@ use App\Http\Controllers\imagecontroller;
 
 Route::middleware(['auth'])->group(function () {
 
-    Route::apiResource("/products", ProductController::class);
+    Route::apiResource("/products", ProductController::class)->middleware([CheckAdmin::class]);
 
-    Route::get('/categories', [ProductController::class, 'getCategories']);
+    Route::get('/categories', [ProductController::class, 'getCategories'])->middleware([CheckAdmin::class]);
 
-    Route::post('/addCategory', [ProductController::class, 'addCategory']);
+    Route::post('/addCategory', [ProductController::class, 'addCategory'])->middleware([CheckAdmin::class]);
 
     // orders routes
     Route::get("{user_id}/orders/range", [OrderController::class,  "get_orders_within_date_range"])->name("orders.range");
     Route::apiResource("{user_id}/orders", OrderController::class);
     // home routes
-    Route::get("/admin/index", [HomeController::class, "admin_index"])->name("admin_index");
+    Route::get("/admin/index", [HomeController::class, "admin_index"])->name("admin_index")->middleware([CheckAdmin::class]);
     Route::get("/rooms", [HomeController::class, "get_rooms"])->name("get_rooms");
     // checks
     Route::get("/orders", [OrderController::class, "get_orders"]);
@@ -41,15 +42,14 @@ Route::middleware(['auth'])->group(function () {
     Route::apiResource("/users", UserController::class);
 
     Route::get("/{user_id}", [HomeController::class, "index"])->name("index");
-    Route::get('/admin/getusers', [admincontroller::class, 'GetAllUsers']);
+    Route::get('/admin/getusers', [admincontroller::class, 'GetAllUsers'])->middleware([CheckAdmin::class]);
 
-    Route::delete('/admin/deleteuser/{id}', [admincontroller::class, 'destroy']);
+    Route::delete('/admin/deleteuser/{id}', [admincontroller::class, 'destroy'])->middleware([CheckAdmin::class]);
 
-    Route::patch('/admin/edituser/{id}', [admincontroller::class, 'update']);
+    Route::patch('/admin/edituser/{id}', [admincontroller::class, 'update'])->middleware([CheckAdmin::class]);
 
     Route::get('/admin/getusers/{id}', [admincontroller::class, 'showuser']);
-    Route::post('/admin/create', [admincontroller::class, 'store']);
-
+    Route::post('/admin/create', [admincontroller::class, 'store'])->middleware([CheckAdmin::class]);
     Route::post('upload', [imagecontroller::class, 'upload']);
     Route::resource('/rooms', roomcontroller::class);
 });
